Add an example DTS for the HiFive Unleashed
This is the device tree that's shipped on the HiFive Unleashed, where it
lives inside the FSBL (which is in a SPI flash).  As that's the only
currently availiable RISC-V Linux board it should serve as a good
reference for people who want to poke around in device tree land.

This device tree contains entries that coorespond to devices on the
MicroSemi expansion board.

Signed-off-by: Palmer Dabbelt <palmer@sifive.com>